feat: e2e 테스트 동작 개선
